iphoneh padding, flex for buttons
Added padding to iphoneh class.  Moved .carti higher up in the CSS. Created a button-container to hold iphone buttons so they could have flex display of row on larger screens and column on small screens. Adjust width of Add Cart button to better match the widths of the iphone buttons.